requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==3.2.15
django-bootstrap5==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==2.1.3
incremental==21.3.0
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.2.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
zipp==3.8.1
zope.interface==5.4.0
